adds branch check to make script safer

+++ update_docs.sh
@@ -10,15 +10,20 @@ tar czf ~/html.tgz .
 cd ../../..
 # checkout doc branch
 git checkout gh-pages
-# clear out old docs
-git rm -rf .
-rm -fr docs
-rm -fr .idea
 
-tar xzf ~/html.tgz
-# commit and push new docs
-git add .
-git commit -a -m "publish the docs"
-git push origin gh-pages
+# make sure the checkout was successful
+current_branch=$(git branch | grep \* | cut -d ' ' -f2-)
+if [ $current_branch = "gh-pages" ]; then
+    # clear out old docs
+    git rm -rf .
+    rm -fr docs
+    rm -fr .idea
 
-git checkout master
+    tar xzf ~/html.tgz
+    # commit and push new docs
+    git add .
+    git commit -a -m "publish the docs"
+    git push origin gh-pages
+
+    git checkout master
+fi
